% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliaNMF.R
\name{nmf_HALS}
\alias{nmf_HALS}
\title{Use Robert Gregg's implementation of HALS for NMF in Julia}
\usage{
nmf_HALS(x, r, alpha = 0.2, maxiter = 250)
}
\arguments{
\item{x}{matrix}

\item{r}{target rank}

\item{alpha}{regularization factor, defaults to 0.2}
}
\value{
a list with elements, X, W, H, res, call
}
\description{
Use Robert Gregg's implementation of HALS for NMF in Julia
}
\examples{
mym = matrix(1:100,nr=10)*1.0
nn = nmf_HALS(mym, 4)
nn$res
rec = nn$W \%*\% t(nn$H)
rec[1:4,1:4]
}
